# docs/swagger.yaml
basePath: /
definitions:
...
  input.AdminDeleteReq:
    properties:
      id:
        description: 管理员ID
        example: 2
        type: integer
    required:
    - id
    type: object
  input.AdminListPageReq:
    properties:
      enable:
        description: 是否启用
        enum:
        - 0
        - 1
        example: 1
        type: integer
      gender:
        description: 性别
        enum:
        - 0
        - 1
        - 2
        example: 1
        type: integer
      pageNo:
        description: 页码
        example: 1
        type: integer
      pageSize:
        description: 每页展示条数
        example: 20
        type: integer
      sortField:
        description: 排序字段
        example: points
        type: string
      sortOrder:
        description: 排序方向 ascend/descend
        enum:
        - ascend
        - descend
        example: descend
        type: string
      username:
        description: 用户名，支持模糊搜索
        example: admin
        type: string
    required:
    - pageNo
    - pageSize
    type: object
...
  resp.MenuValidateResp:
    properties:
      has:
        description: 是否存在
        example: true
        type: boolean
    type: object
...
  response.Response:
    properties:
      code:
        description: 响应code
        example: 0
        type: integer
      data:
        description: 响应数据
      msg:
        description: 响应信息
        example: success
        type: string
    type: object
info:
  contact:
    email: junjie.he.925@gmail.com
    name: 何俊杰
  description: BiliLive Assistant 系统接口文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: BiliLive Assistant API
  version: "1.0"
paths:
  /api/admin/admin/delete:
    post:
      description: 根据管理员 ID 删除管理员账号（通常不允许删除当前登录管理员）
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 删除参数（管理员 ID）
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.AdminDeleteReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除管理员
      tags:
      - 管理员
  /api/admin/admin/list:
    post:
      description: 分页获取管理员列表，支持按用户名、性别、启用状态进行筛选
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.AdminListPageReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.AdminListPageResp'
              type: object
      security:
      - BearerAuth: []
      summary: 分页查询管理员列表
      tags:
      - 管理员
  /api/admin/admin/save:
    post:
      description: 根据是否传入 ID 创建或更新管理员账号：未传 ID 时为创建，传入 ID 时为更新，同时可设置账号状态及绑定角色
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 管理员保存参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.AdminSaveReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建或更新管理员
      tags:
      - 管理员
  /api/admin/admin/update-password:
    post:
      description: 管理员根据管理员 ID 重置指定账号的登录密码（无需旧密码，通常需要具备管理员管理权限）
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 重置密码参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.AdminResetAdminPasswordReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 重置管理员密码
      tags:
      - 管理员
  /api/admin/admin/update-profile:
    post:
      description: 管理员更新个人资料信息（如昵称、头像、性别、邮箱等），支持部分字段更新
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 管理员个人资料参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.AdminUpdateProfileReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新管理员个人资料
      tags:
      - 管理员
  /api/admin/appconfig/data:
    post:
      description: 获取后台可编辑的全部 App 配置（站点信息、颜色、图标、注册开关、登录页配置），用于回填管理页表单
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.AppConfigDataResp'
              type: object
      security:
      - BearerAuth: []
      summary: 获取 App 配置
      tags:
      - App配置
  /api/admin/appconfig/oss_save:
    post:
      description: 整体覆盖保存 OSS 相关配置并刷新缓存立即生效，可为空表示清除对应配置
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: App 配置参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.AppConfigSaveOssConfigReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 保存 OSS 配置
      tags:
      - App配置
  /api/admin/appconfig/save:
    post:
      description: 整体覆盖保存基本 App 配置并刷新缓存立即生效，可为空表示清除对应配置
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: App 配置参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.AppConfigSaveConfigReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 保存 App 基础配置
      tags:
      - App配置
  /api/admin/auth/captcha:
    post:
      description: 获取当前系统是否已配置并启用了 altcha 验证码，前端可根据此状态决定是否展示验证码组件
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.AdminCaptchaResp'
              type: object
      summary: 获取验证码启用状态
      tags:
      - 认证
  /api/admin/auth/change-password:
    post:
      description: 管理员在已登录状态下通过旧密码验证后修改登录密码，修改成功后建议重新登录以获取新的凭证
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 修改密码参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.AdminChangePasswordReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 修改登录密码
      tags:
      - 认证
  /api/admin/auth/detail:
    post:
      description: 获取当前登录管理员的详细信息，包括基础信息、个人资料、拥有角色列表及当前角色信息
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.AdminDetailsResp'
              type: object
      security:
      - BearerAuth: []
      summary: 获取当前管理员信息
      tags:
      - 认证
  /api/admin/auth/login:
    post:
      description: 管理员通过账号和密码登录系统，登录成功后返回 access_token 和 refresh_token，用于后续接口鉴权
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 登录参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.AdminLoginReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.AdminLoginResp'
              type: object
      summary: 管理员登录
      tags:
      - 认证
  /api/admin/auth/logout:
    post:
      description: 管理员退出当前登录状态，使当前 access_token 失效（通常用于前端主动登出）
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 管理员退出登录
      tags:
      - 认证
  /api/admin/auth/refresh:
    post:
      description: 使用 refresh_token 刷新登录状态，获取新的 access_token 和 refresh_token，用于延长会话有效期
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 刷新凭证参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.AdminRefreshReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.AdminLoginResp'
              type: object
      summary: 刷新登录凭证
      tags:
      - 认证
  /api/admin/auth/switch-role:
    post:
      description: 管理员在已登录状态下切换当前角色，根据目标角色重新生成 access_token 和 refresh_token，用于更新权限上下文
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 切换角色参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.AdminSwitchRoleReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.AdminLoginResp'
              type: object
      security:
      - BearerAuth: []
      summary: 切换管理员角色
      tags:
      - 认证
  /api/admin/feedback/details:
    post:
      description: 根据投诉ID返回投诉全部字段（含投诉正文），以及投诉用户的 uid/uname
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.FeedbackDetailsReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.FeedbackDetailsResp'
              type: object
      security:
      - BearerAuth: []
      summary: 后台获取投诉详情
      tags:
      - 投诉管理
  /api/admin/feedback/list:
    post:
      description: 分页查询用户提交的投诉/反馈，联查用户表返回用户 UID 与昵称；支持按 UID 精确、按昵称模糊筛选。列表不下发投诉正文，正文请调详情接口
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.FeedbackListPageReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.FeedbackListPageResp'
              type: object
      security:
      - BearerAuth: []
      summary: 后台分页查询投诉列表
      tags:
      - 投诉管理
...
  /api/admin/menu/buttons:
    post:
      description: 根据菜单 ID 获取其下的按钮权限列表（按钮类型菜单），用于前端页面按钮级权限控制
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 查询参数（父级菜单 ID）
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.MenuButtonsReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.MenuButtonsResp'
              type: object
      security:
      - BearerAuth: []
      summary: 获取菜单下的按钮权限
      tags:
      - 菜单
  /api/admin/menu/delete:
    post:
      description: 根据菜单 ID 删除菜单（通常不允许删除存在子菜单或已绑定权限的菜单）
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 删除参数（菜单 ID）
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.MenuDeleteReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除菜单
      tags:
      - 菜单
  /api/admin/menu/list:
    post:
      description: 获取系统完整菜单树结构（包含菜单及按钮权限），用于前端动态路由生成和权限控制
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.MenuListResp'
              type: object
      security:
      - BearerAuth: []
      summary: 获取菜单列表
      tags:
      - 菜单
  /api/admin/menu/save:
    post:
      description: 根据是否传入 ID 创建或更新菜单信息：未传 ID 时为创建，传入 ID 时为更新，支持菜单及按钮类型（MENU / BUTTON），用于构建系统菜单结构和权限控制
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 菜单保存参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.MenuSaveReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建或更新菜单
      tags:
      - 菜单
  /api/admin/menu/toggle:
    post:
      description: 根据菜单 ID 切换菜单的启用状态（启用/禁用），常用于后台列表中的快速开关操作
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 切换参数（菜单 ID）
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.MenuToggleReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 切换菜单启用状态
      tags:
      - 菜单
  /api/admin/menu/validate:
    post:
      description: 根据菜单路径校验菜单是否已存在，常用于前端需要对菜单进行校验时
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 校验参数（菜单路径）
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.MenuValidateReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.MenuValidateResp'
              type: object
      security:
      - BearerAuth: []
      summary: 校验菜单路径是否存在
      tags:
      - 菜单
...
  /auth/altcha/challenge:
    get:
      description: 生成 altcha 工作量证明挑战，供前端验证码组件使用。未配置 hmacKey（即未启用验证码）或生成失败时返回空对象 {}
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      responses:
        "200":
          description: altcha 挑战对象；未启用验证码或生成失败时为空对象 {}
          schema:
            type: object
      summary: 获取 altcha 验证码挑战
      tags:
      - 通用
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"